+++ test/features/multiple_one_of/multiple_one_of_test.rb
@@ -2,12 +2,12 @@
 
 require "test_helper"
 
-require_relative "../support/examples/card_payment_component"
-require_relative "../support/examples/bank_payment_component"
-require_relative "../support/examples/standard_shipping_component"
-require_relative "../support/examples/express_shipping_component"
-require_relative "../support/examples/with_id_component"
-require_relative "../support/examples/without_id_component"
+require_relative "../../support/examples/card_payment_component"
+require_relative "../../support/examples/bank_payment_component"
+require_relative "../../support/examples/standard_shipping_component"
+require_relative "../../support/examples/express_shipping_component"
+require_relative "../../support/examples/with_id_component"
+require_relative "../../support/examples/without_id_component"
 
 # Tests for multiple named oneOf properties in a single version
 class Verquest::MultipleOneOfTest < Minitest::Test
@@ -205,161 +205,5 @@ class Verquest::MultipleOneOfTest < Minitest::Test
 
   def test_valid_schema
     assert OrderWithPaymentAndShipping.valid_schema?(version: "2025-06")
-  end
-end
-
-# Tests for multiple oneOf without discriminators (schema inference)
-class Verquest::MultipleOneOfWithoutDiscriminatorTest < Minitest::Test
-  class MultiContentRequest < Verquest::Base
-    version "2025-06" do
-      field :request_id, type: :string, required: true
-
-      one_of name: :primary_content do
-        reference :with_id, from: WithIdComponent
-        reference :without_id, from: WithoutIdComponent
-      end
-
-      one_of name: :secondary_content do
-        reference :with_id, from: WithIdComponent
-        reference :without_id, from: WithoutIdComponent
-      end
-    end
-  end
-
-  def test_mapping_one_ofs_count
-    mapping = MultiContentRequest.mapping(version: "2025-06")
-
-    assert_equal 2, mapping["_oneOfs"].size
-  end
-
-  def test_mapping_all_one_ofs_have_variant_schemas
-    mapping = MultiContentRequest.mapping(version: "2025-06")
-
-    mapping["_oneOfs"].each do |one_of_mapping|
-      assert one_of_mapping.key?("_variant_schemas"), "Expected _variant_schemas for schema inference"
-      assert one_of_mapping.key?("_variant_path")
-    end
-  end
-
-  def test_process_both_with_id
-    params = {
-      "request_id" => "req-001",
-      "primary_content" => {
-        "id" => "primary-1",
-        "name" => "Primary Content",
-        "value" => 100
-      },
-      "secondary_content" => {
-        "id" => "secondary-1",
-        "name" => "Secondary Content",
-        "value" => 200
-      }
-    }
-    Verquest.configuration.validation_error_handling = :result
-
-    result = MultiContentRequest.process(params, version: "2025-06", validate: true)
-
-    assert_predicate result, :success?
-    assert_equal params, result.value
-  ensure
-    Verquest.configuration.validation_error_handling = :raise
-  end
-
-  def test_process_mixed_variants
-    params = {
-      "request_id" => "req-002",
-      "primary_content" => {
-        "id" => "primary-1",
-        "name" => "Primary Content",
-        "value" => 100
-      },
-      "secondary_content" => {
-        "name" => "Secondary Content",
-        "description" => "No ID here"
-      }
-    }
-    Verquest.configuration.validation_error_handling = :result
-
-    result = MultiContentRequest.process(params, version: "2025-06", validate: true)
-
-    assert_predicate result, :success?
-    assert_equal params, result.value
-  ensure
-    Verquest.configuration.validation_error_handling = :raise
-  end
-
-  def test_valid_schema
-    assert MultiContentRequest.valid_schema?(version: "2025-06")
-  end
-end
-
-# Tests for mixed discriminator and schema-inference oneOf
-class Verquest::MixedMultipleOneOfTest < Minitest::Test
-  class MixedRequest < Verquest::Base
-    version "2025-06" do
-      field :id, type: :string, required: true
-
-      # OneOf with discriminator
-      one_of name: :payment, discriminator: :method, required: true do
-        reference :card, from: CardPaymentComponent
-        reference :bank, from: BankPaymentComponent
-      end
-
-      # OneOf without discriminator (schema inference)
-      one_of name: :metadata do
-        reference :with_id, from: WithIdComponent
-        reference :without_id, from: WithoutIdComponent
-      end
-    end
-  end
-
-  def test_mapping_one_ofs_count
-    mapping = MixedRequest.mapping(version: "2025-06")
-
-    assert_equal 2, mapping["_oneOfs"].size
-  end
-
-  def test_mapping_payment_one_of_with_discriminator
-    mapping = MixedRequest.mapping(version: "2025-06")
-    payment_mapping = mapping["_oneOfs"].find { |m| m.key?("_discriminator") }
-
-    assert payment_mapping, "Expected payment oneOf with discriminator"
-    assert_equal "payment/method", payment_mapping["_discriminator"]
-  end
-
-  def test_mapping_metadata_one_of_with_schema_inference
-    mapping = MixedRequest.mapping(version: "2025-06")
-    metadata_mapping = mapping["_oneOfs"].find { |m| m.key?("_variant_schemas") }
-
-    assert metadata_mapping, "Expected metadata oneOf with variant schemas"
-    assert_equal "metadata", metadata_mapping["_variant_path"]
-  end
-
-  def test_process_with_discriminator_and_schema_inference
-    params = {
-      "id" => "mixed-001",
-      "payment" => {
-        "method" => "card",
-        "card_number" => "4111111111111111",
-        "expiry" => "12/25"
-      },
-      "metadata" => {
-        "id" => "meta-1",
-        "name" => "Metadata with ID",
-        "value" => 42
-      }
-    }
-    Verquest.configuration.validation_error_handling = :result
-
-    result = MixedRequest.process(params, version: "2025-06", validate: true)
-
-    assert_predicate result, :success?
-    assert_equal params, result.value
-  ensure
-    Verquest.configuration.validation_error_handling = :raise
-  end
-
-  def test_valid_schema
-    assert MixedRequest.valid_schema?(version: "2025-06")
   end
 end
